# Explicit ordering of Content Items to ensure deterministic HTTP responses